requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
colorama==0.4.6
decorator==5.1.1
Django==4.2.6
django-bootstrap-v5==1.0.11
django-extensions==3.2.3
django-appconf==1.0.5
django-imagekit==5.0.0
exceptiongroup==1.1.3
executing==2.0.0
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
pilkit==3.0
